Chore: fix CSS reset, move Sandbox components into own file
