+++ sandbox/index.html
@@ -1,10 +1,6 @@
 <html>
   <head>
     <title>Component lib sandbox</title>
-    <!-- <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/meyer-reset/2.0/reset.min.css"
-    /> -->
   </head>
   <body>
     <div id="root"></div>
